status variable should be defined as DWORD to be properly stored inside registry

+++ mDNSWindows/mDNSWin32.c
@@ -1653,7 +1653,7 @@ mDNSexport mDNSBool mDNSPlatformSetDNSConfig(mDNSBool setservers, mDNSBool setse
 mDNSexport void mDNSPlatformDynDNSHostNameStatusChanged(const domainname *const dname, const mStatus status)
 {
 	char		uname[MAX_ESCAPED_DOMAIN_NAME];
-	BYTE		bStatus;
+	DWORD		bStatus;
 	LPCTSTR		name;
 	HKEY		key = NULL;
 	mStatus		err;
